setup: Support Zsh completions for setups with ZDOTDIR paths
Some users may like to follow the XDG Base Directory specification, and
Zsh partially supports it as it allows the user to define environmental
variables, such as ZDOTDIR, to override the default locations of dot
files.

With this commit, when there is no ~/.zshrc the setup script will now
check if user has the ZDOTDIR directory defined and then if there is
a .zshrc file inside. If it does, then the kw native Zsh completions
necessary configs will be added to this file.

Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: Lincoln Yuji <lincolnyuji@hotmail.com>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>