setup: Allow ZDOTDIR paths for zsh autocompletion setups
Some users may like to follow the XDG Base Directory specification,
and zsh partially supports it as it allows the user to define
environmental variables, such as ZDOTDIR, to override the default
locations of dot files.

With this update, when there is no ~/.zshrc the setup script will now
check if user has the ZDOTDIR directory defined and then if there is a
.zshrc file inside. If it does, then the auto complete will be
installed on this file.

Signed-off-by: Lincoln Yuji <lincolnyuji@hotmail.com>