src: plugins: kernel_install: Add force option for deploy
When deploying a kernel to some Arch Linux-based distro with mkinitcpio,
users can see some non-fatal error, i.e., the error does not prevent the
initramfs from being generated. Those cases might be harmless, but
asking the user if they are okay with proceeding with those errors is
better. For this reason, this commit introduces an interactive prompt
with the user to see if they want to proceed. This commit also adds the
force option to skip any interaction with the prompt.

Closes: #1083

Reviewed-by: Melissa Wen <melissa.srw@gmail.com>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>