src: plugins: kernel_uninstall: utils: Move INSTALLED_KERNELS to /boot
Distros that adopt an A/B partition system usually replace the entire
/boot folder but preserve the content in the /opt folder, creating
inconsistencies with kw after a distro update since the kw list will
point to kernels that no longer exist. To handle this case, this commit
moves the INSTALLED_KERNEL list to the /boot folder, but some old
versions of the kw will be affected by this change; for this reason,
this function migrates those old files to the new scheme. At some point,
we can safely remove this migration function.

Closes: #886

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>