src: plugins: kernel_install: utils: Adjust deploy list for RO /boot
In an A/B system that has an RO /boot and no INSTALLED_KERNELS files,
users can see an error when trying to list kernels. This commit
addresses this issue by checking if the boot is read-only and the
INSTALLED_KERNELS file is unavailable, which avoids displaying system
errors to the users.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>