.github: workflows: Remove kcov job
I knew for a long time that the Kcov package was unavailable in the
latest Debian versions, but I kept the support for Kcov since CI did not
break. However, the latest version of Ubuntu available in the CI does
not have kcov anymore. Right now, I don't have enough time to fix this
issue, and I consider this topic a low priority. For all of those
reasons, this commit disables the kcov job.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>