src: remote: Fix bug when the user asks to remove IP from known host
When the target IP is already in the know host file, kw pop up a
message to users asking them if they want to remove the target IP from
the file.  This operation failed due to a mistake in how kw handles the
IP information to compose the ssh-keygen command; this commit addressed
this issue using proper syntax.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>