src: plugins: kernel_install: rpi_bootloader: Add missing kernel image name
The API for run_bootloader_update has four parameters, and the last one
is the kernel binary file name. For some reason, run_bootloader_update
for RPI was not using that information to identify the kernel binary,
which might lead to errors. This commits adds the missing parameter,
followed by the necessary adaptation to the unit test.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>