src: deploy: Send all files under kernel_install path
A long time ago, we added a simple optimization to only send required
files to the target machine instead of sending everything. This
optimization does not provide any substantial gain and complicates code
maintenance. For this reason, this commit changes the deploy approach to
send everything from the kernel_install to the target machine.

Signed-off-by: Rodrigo Siqueira <siqueira@igalia.com>