setup: port kw for MacOS
Some kernel developers use a MacOS host as a development environment to
leverage the Apple Silicon ARM chips to build for Raspberry Pi, for
example. kw can support this context as it is entirely written in Bash,
and we mostly need to map dependencies from Homebrew [1].

To support MacOS, change the `setup.sh` script to detect a MacOS system
and install dependencies using Homebrew. In `setup.sh`, the adaptation
needed was only to detect MacOS using `uname --kernel-name` and direct
the correct `brew` commands. The list of base dependencies (at
`documentation/dependencies`) is minimal to packages that seem core,
while the build dependencies (at
`documentation/dependencies/kernel_build`) are empty.

[NOTE] I tested this in an "old" 2016 MacBookPro and I am unable (for
some reason) to update to a MacOS version supported by Homebrew, which
adverts that it is expected that "formulae will fail to build" and that
"Homebrew will be slow and buggy". Due to this, I couldn't install most
packages and essentially tested if the `setup.sh` script was reliable in
detecting MacOS and running the correct `brew` commands.

[1]: https://brew.sh/

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>