src: deploy: Make kw uninstall parameter optional
To remove a kernel, the user must first run "kw deploy --list" and take
the kernel name from the list. This commit enables the "--uninstall"
option to work without a parameter, which makes kw remove the first
managed kernel it encounters, which is especially useful when the user
only has one kernel installed.

Closes: #1192

Suggested-by: Mario Limonciello <mario.limonciello@amd.com>
Signed-off-by: Rodrigo Siqueira <siqueira@igalia.com>