src: tag: Add new feature 'kw tag'
Often a kernel developer has to perform some operations over
commit or patch trailer lines. Adding tags, such as 'Reviewed-by',
is a common task and sometimes it might be required to do this over
multiple commits or patches. Also another usual one is the 'Fixes'
tag, that must follow a specific format.

This feature helps on these cases. It uses 'git rebase',
'git commit --amend' and 'git interpret-trailers' as its
backend in order to make these tasks quicker to complete.

Co-authored-by: Luiza Soezima <lbrsoezima@usp.br>
Co-authored-by: Sabrina Araujo <sabrinaaraujo@usp.br>
Signed-off-by: Lincoln Yuji <lincolnyuji@hotmail.com>
Signed-off-by: Bruna Bispo <blbispo1@gmail.com>