src: build_and_deploy: make kw bd accept kw d arguments
Previously, to build and deploy the kernel, users had to use kw bd, but it did not accept the same arguments as kw d, which was inconsistent and unintuitive.

this refactor ensures that kw bd now accepts the same arguments as kw d, providing a more coherent and flexible interface for users.

 Signed-off-by: Nattan Ferreira <nattanferreira58@gmail.com>
 Co-developed-by: Lucas Antonio <lucasantonio.santos@usp.br>
 Signed-off-by: Lucas Antonio <lucasantonio.santos@usp.br>