src: plugins: kernel_install: Run kernel-install in some cases
When using systemd-boot, the kernel-install command is usually invoked
by some of the distro update scripts. Unfortunately, all distros handle
it slightly differently, and in some cases, the kernel-install command
is not invoked, which can result in a deployed kernel that does not show
up in systemd-boot. This commit addresses this issue by invoking
kernel-install if the entry for the new kernel is not available. The
drawback of this approach is that the initrd might be generated a second
time, but this does not cause any issues.

Signed-off-by: Rodrigo Siqueira <siqueira@igalia.com>