src: plugin: kernel_install: Add PopOS workaround
PopOS uses Kernelstub instead of kernel-install. Additionally, this OS
features a partition system with A and B partitions, where any newly
deployed kernel becomes the default kernel, and the old kernel remains
in the boot system. The problem with this approach is that multiple
deploys will eventually cause the distro kernel to be lost from the
boot. I did not find any good solution that respects the PopOS way of
doing things. To mitigate part of this problem, kw still uses
kernel-install in PopOS (even though this is not the recommended way)
and does one manual change to the entry file to make the new kernel
visible in the systemd-boot menu and reaffirm the latest generic kernel.
This approach does not mess with the A and B partition systems used by
PopOs, but it is not a long term solution.

Signed-off-by: Rodrigo Siqueira <siqueira@igalia.com>