src: plugins: kernel_install: systemd_boot: Avoid match collision
At some point, I wrote a poor string manipulation based on the cut
command. The idea was to extract the kernel name after vmlinuz, for
example:

/boot/vmlinuz-6.12.10-76061203-generic

The idea was to collect 6.12.10-76061203-generic. Everything was working
fine until I started to see some flake test results that failed randomly
after a couple of executions. After one hour of debugging, I figured out
that the issue originated from the temporary folders generated during
every execution; whenever the folder contained the letter 'z', our cut
command took action and corrupted the result. For example, if the find
command returned:

/boot/lazblaz/vmlinuz-6.12.10-76061203-generic
        ^

Our cut will take action and return "bla" which is wrong. To address
this issue, I decided to use a regex with a very clever tool from Perl,
which is the \K option (for more details, check
https://perldoc.perl.org/perlre#%5CK).

Technically, this commit should never exist since the original set of
changes was not merged. However, this problem is so interesting that I
felt it was worth registering as a commit for educational purposes and
also as an alert not to use 'cut' every time.

Signed-off-by: Rodrigo Siqueira <siqueira@igalia.com>