src: plugins: kernel_install: rpi_bootloader: add overlay copy
During the compilation of a downstream rpi kernel some overlays file are
created together with dtb files. kw is currently copying those files and
the overlays directory to the kw package for remote deployment, but it's
not copying overlays from kw package to the remote target directory.
This commit adds this missing step.

Co-developed-by: Rodrigo Siqueira <siqueira@igalia.com>
Signed-off-by: Melissa Wen <mwen@igalia.com>