src: transition_functions: Introduce transition file
When dealing with updates in kw, sometimes we have to implement
temporary functions to migrate some old way of doing things to a modern
approach. This commit introduces a file that should centralize these
types of functions. The first function introduced is the one that
migrates the old env folder scheme to the new one.

Signed-off-by: Rodrigo Siqueira <siqueira@igalia.com>